Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main (){
-    int a,s,d;
-    scanf("%d%d%d",&a,&s,&d );
+    float a,s,d;
+    scanf("%f%f%f",&a,&s,&d );
     int n= (a+s+d)/3;
     printf("%.2f",n);
     return 0;
